+++ driver/src/main/java/oracle/nosql/driver/iam/FederationRequestHelper.java
@@ -23,8 +23,6 @@ import java.util.Date;
 import java.util.Set;
 import java.util.logging.Logger;
 
-import javax.net.ssl.SSLException;
-
 import oracle.nosql.driver.httpclient.HttpClient;
 import oracle.nosql.driver.iam.CertificateSupplier.X509CertificateKeyPair;
 import oracle.nosql.driver.util.HttpRequestUtil;
@@ -35,8 +33,6 @@ import com.fasterxml.jackson.core.JsonParser;
 
 import io.netty.handler.codec.http.DefaultHttpHeaders;
 import io.netty.handler.codec.http.HttpHeaders;
-import io.netty.handler.ssl.SslContext;
-import io.netty.handler.ssl.SslContextBuilder;
 
 /**
  * @hidden
@@ -49,71 +45,34 @@ class FederationRequestHelper {
     private static final String APP_JSON = "application/json";
     private static final String DEFAULT_FINGERPRINT = "SHA256";
 
-    static String getSecurityToken(URI endpoint,
+    static String getSecurityToken(HttpClient client,
+                                   URI endpoint,
                                    int timeoutMs,
                                    String tenantId,
                                    X509CertificateKeyPair pair,
                                    String body,
                                    Logger logger) {
+        CharBuffer charBuf = CharBuffer.wrap(body);
+        ByteBuffer buf = StandardCharsets.UTF_8.encode(charBuf);
+        byte[] payloadByte = new byte[buf.remaining()];
+        buf.get(payloadByte);
 
-        HttpClient client = null;
-        try {
-            client = buildHttpClient(endpoint, logger);
-            CharBuffer charBuf = CharBuffer.wrap(body);
-            ByteBuffer buf = StandardCharsets.UTF_8.encode(charBuf);
-            byte[] payloadByte = new byte[buf.remaining()];
-            buf.get(payloadByte);
+        HttpResponse response = HttpRequestUtil.doPostRequest(
+            client, endpoint.toString(),
+            headers(tenantId, endpoint, payloadByte, pair, logger),
+            payloadByte, timeoutMs, logger);
 
-            HttpResponse response = HttpRequestUtil.doPostRequest(
-                client, endpoint.toString(),
-                headers(tenantId, endpoint, payloadByte, pair, logger),
-                payloadByte, timeoutMs, logger);
-
-            int responseCode = response.getStatusCode();
-            if (responseCode > 299) {
-                throw new IllegalStateException(
-                    String.format(
-                        "Error getting security token from IAM, " +
-                        "status code %d, response \n%s",
-                        response.getStatusCode(),
-                        response.getOutput()));
-            }
-            logTrace(logger, "Federation response " + response.getOutput());
-            return parseResponse(response.getOutput());
-        } finally {
-            if (client != null) {
-                client.shutdown();
-            }
-        }
-    }
-
-    private static HttpClient buildHttpClient(URI endpoint, Logger logger) {
-        String scheme = endpoint.getScheme();
-        if (scheme == null) {
-            throw new IllegalArgumentException(
-                "Unable to find URL scheme, invalid URL " +
-                 endpoint.toString());
-        }
-        if (scheme.equalsIgnoreCase("http")) {
-            return HttpClient.createMinimalClient(endpoint.getHost(),
-                                                  endpoint.getPort(),
-                                                  null,
-                                                  "FederationClient",
-                                                  logger);
-        }
-
-        SslContext sslCtx = null;
-        try {
-            sslCtx = SslContextBuilder.forClient().build();
-        } catch (SSLException se) {
+        int responseCode = response.getStatusCode();
+        if (responseCode > 299) {
             throw new IllegalStateException(
-                "Unable to build SSL context for http client", se);
+                String.format(
+                    "Error getting security token from IAM, " +
+                    "status code %d, response \n%s",
+                    response.getStatusCode(),
+                    response.getOutput()));
         }
-        return HttpClient.createMinimalClient(endpoint.getHost(),
-                                              443,
-                                              sslCtx,
-                                              "FederationClient",
-                                              logger);
+        logTrace(logger, "Federation response " + response.getOutput());
+        return parseResponse(response.getOutput());
     }
 
     /*